requirements.txt
# torch==2.1.0+cu126
# torchvision==0.15.2+cu126
Pillow>=8.0.0
# numpy==1.21.2
opencv-python==4.5.5.64
zipfile36==0.1.3
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0